Export less from the server lib
This exports run_all instead of run_shared_state and
run_on_new_localset, which lets us hide a lot of types.
